Changing build system to Scons, fixes for Linux x64
-- The most major change of this commit is that the Refu Builder is retired from the Public branch of Refu and SCons is replacing it as the build system
-- Many bug fixes thanks to more extensive testing
-- Fixed bugs specific to some x64 Linux distros
-- Lots of other smaller changes that were performed over a long time period in the Private branch of Refu

+++ src/String/manipulation.c
@@ -365,7 +365,7 @@ char i_rfString_Replace(RF_String* thisstr,const void* sstrP,const void* rstrP,c
     RF_MALLOC(bytePositions,bSize*sizeof(int32_t));
     //if the given num is 0 just make sure we replace all
     if(number == 0)
-        number = ULONG_MAX;//max number of occurences
+        number = UINT_MAX;//max number of occurences
 
     //find how many occurences exist
     rfStringX_FromString_IN(&temp,thisstr);
